LayoutTests/fast/text-extraction/debug-text-extraction-interactions.html

<!-- webkit-test-runner [ useFlexibleViewport=true textExtractionEnabled=true ] -->
<!DOCTYPE html>
<html>
<meta name="viewport" content="width=device-width, initial-scale=1">
<meta charset="utf-8">
<head>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
</head>
<body>
    <button id="test-button" aria-label="Click Me">Test</button>
    <input type="email" placeholder="Enter some text" value="foo" />
    <select role="menu">
        <option>One</option>
        <option selected>Two</option>
        <option>Three</option>
    </select>
    <div contenteditable="plaintext-only">
        <h3 aria-label="Heading">Subject</h3>
        <p>Hello world.</p>
    </div>
    <h1 class="click-count">0</h1>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        textField = document.querySelector("input");
        select = document.querySelector("select");
        clickCount = document.querySelector(".click-count");
        document.getElementById("test-button").addEventListener("click", () => {
            clickCount.textContent = parseInt(clickCount.textContent) + 1;
        });

        if (!window.testRunner)
            return;

        const debugText = await UIHelper.requestDebugText();

        clickSucceeded = await UIHelper.performTextExtractionInteraction("click", {
            nodeIdentifier: UIHelper.nodeIdentifierFromDebugText(debugText, "Click Me")
        });

        textInputSucceeded = await UIHelper.performTextExtractionInteraction("textinput", {
            nodeIdentifier: UIHelper.nodeIdentifierFromDebugText(debugText, "Enter some text"),
            replaceAll: true,
            text: "bar"
        });

        selectMenuItemSucceeded = await UIHelper.performTextExtractionInteraction("selectmenuitem", {
            nodeIdentifier: UIHelper.nodeIdentifierFromDebugText(debugText, "menu"),
            text: "Three"
        });

        selectTextSucceeded = await UIHelper.performTextExtractionInteraction("selecttext", {
            nodeIdentifier: UIHelper.nodeIdentifierFromDebugText(debugText, "contentEditable"),
            text: "Subject"
        });

        invalidActionSucceeded = await UIHelper.performTextExtractionInteraction("click", {
            nodeIdentifier: "invalid_id"
        });

        shouldBeTrue("clickSucceeded");
        shouldBeEqualToString("clickCount.textContent", "1");
        shouldBeTrue("textInputSucceeded");
        shouldBeEqualToString("textField.value", "bar");
        shouldBeTrue("selectMenuItemSucceeded");
        shouldBeEqualToString("select.value", "Three");
        shouldBeTrue("selectTextSucceeded");
        shouldBeEqualToString("getSelection().toString()", "Subject");
        shouldBeFalse("invalidActionSucceeded");

        finishJSTest();
    });
    </script>
</body>
</html>
